# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yavabali API",
        "description": "API untuk mengelola data bahan makanan",
        "version": "1.0"
    },
    "paths": {
        "/api/ingredients": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Get all ingredients",
                "operationId": "88554df920fe97e07f179ca559f98b75",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Create new ingredient",
                "operationId": "290f4795ec7faed0d1e88d61aeb935ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "NAME"
                                ],
                                "properties": {
                                    "NAME": {
                                        "type": "string"
                                    },
                                    "origin": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "nutritional_info": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "properties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ingredient created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/ingredients/{slug}": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Show ingredient by slug",
                "operationId": "bc6521d52642f36fdbd7d2dcf943fa4f",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Ingredient not found"
                    }
                }
            }
        },
        "/api/ingredients/{id}": {
            "put": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Update ingredient",
                "operationId": "b2b705f67b293905971309a7927b3636",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "NAME": {
                                        "type": "string"
                                    },
                                    "origin": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "nutritional_info": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "properties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ingredient updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Delete ingredient",
                "operationId": "d7a227e0854bd498dfe1c1b9155b4bd0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ingredient deleted successfully"
                    },
                    "404": {
                        "description": "Ingredient not found"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "summary": "Get all products",
                "operationId": "24c51a4469fbe1d6ef9c4d258e31bbf3",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Ingredients",
            "description": "API terkait data bahan makanan"
        }
    ]
}